+++ scripts/update-oembed.php
@@ -24,6 +24,7 @@ saveEndpoint('https://api.crowdsignal.com/oembed', ['polldaddy.com/poll/*', 'pol
 saveEndpoint('http://www.scribd.com/services/oembed/', ['www.scribd.com/document/*']);
 saveEndpoint('https://api.imgur.com/oembed', ['imgur.com/*', 'i.imgur.com/*']);
 saveEndpoint('https://www.youtube.com/oembed', ['*.youtube.com/playlist?*']);
+saveEndpoint('https://www.tiktok.com/oembed', ['*.tiktok.com/*']);
 
 //Export the endpoints
 file_put_contents(