+++ tests/fixtures/www.tiktok.com.a3noticias-video-6806030056956251397.php
@@ -1,0 +1,142 @@
+<?php
+declare(strict_types = 1);
+
+return [
+    'authorName' => 'Antena 3 Noticias',
+    'authorUrl' => 'https://www.tiktok.com/@a3noticias',
+    'cms' => null,
+    'code' => [
+        'html' => '<blockquote class="tiktok-embed" cite="https://www.tiktok.com/@a3noticias/video/6806030056956251397" data-video-id="6806030056956251397" style="max-width: 605px;min-width: 325px;" > <section> <a target="_blank" title="@a3noticias" href="https://www.tiktok.com/@a3noticias">@a3noticias</a> <p>Nuestro <a title="aplausosanitario" target="_blank" href="https://www.tiktok.com/tag/aplausosanitario">#AplausoSanitario</a> más sincero. Para todos. Y recuerda, <a title="quédateencasa" target="_blank" href="https://www.tiktok.com/tag/quédateencasa">#QuédateEnCasa</a></p> <a target="_blank" title="♬ Resistiré - Duo Dinamico" href="https://www.tiktok.com/music/Resistiré-250824169968246784">♬ Resistiré - Duo Dinamico</a> </section> </blockquote> <script async src="https://www.tiktok.com/embed.js"></script>',
+        'width' => 100,
+        'height' => 100
+    ],
+    'description' => 'Nuestro #AplausoSanitario más sincero. Para todos. Y recuerda, #QuédateEnCasa',
+    'favicon' => 'https://s16.tiktokcdn.com/musical/resource/wap/static/image/logo_144c91a.png?v=2',
+    'feeds' => [],
+    'icon' => 'https://s16.tiktokcdn.com/musical/resource/mtact/static/pwa/icon_128x128.png',
+    'image' => 'https://p16-va-default.akamaized.net/obj/tos-maliva-p-0068/26093075d77684fa3d60fd42bf50d48a',
+    'keywords' => [
+        'a3noticias',
+        'antena 3 noticias',
+        'aplausosanitario',
+        'quédateencasa'
+    ],
+    'language' => 'en',
+    'languages' => [
+        'x-default' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397',
+        'id' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=id',
+        'de' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=de',
+        'en' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=en',
+        'es' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=es',
+        'fr' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=fr',
+        'it' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=it',
+        'pl' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=pl',
+        'pt-BR' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=pt_BR',
+        'vi' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=vi',
+        'tr' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=tr',
+        'ru' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=ru',
+        'hi-IN' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=hi',
+        'ko' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=ko',
+        'ja' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=ja',
+        'zh-Hant' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=zh_Hant',
+        'ar' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397?lang=ar'
+    ],
+    'license' => null,
+    'providerName' => 'TikTok',
+    'providerUrl' => 'https://www.tiktok.com/',
+    'publishedTime' => null,
+    'redirect' => null,
+    'title' => 'Nuestro #AplausoSanitario más sincero. Para todos. Y recuerda, #QuédateEnCasa',
+    'url' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397',
+    'linkedData' => [
+        '@context' => 'http://schema.org/',
+        '@type' => 'VideoObject',
+        'url' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397',
+        'name' => 'Antena 3 Noticias(@a3noticias) on TikTok Nuestro #AplausoSanitario más sincero. Para todos. Y recuerda, #QuédateEnCasa',
+        'description' => 'Antena 3 Noticias(@a3noticias) has created a short video on TikTok with music Resistiré. Nuestro #AplausoSanitario más sincero. Para todos. Y recuerda, #QuédateEnCasa',
+        'keywords' => 'a3noticias, Antena 3 Noticias,aplausosanitario,quédateencasa',
+        'thumbnailUrl' => [
+            'https://p16-va-default.akamaized.net/obj/tos-maliva-p-0068/26093075d77684fa3d60fd42bf50d48a',
+            'https://p16-va-default.akamaized.net/obj/tos-maliva-p-0068/f6341ac3dbb54a188f2ea94ac27dd848_1584652366'
+        ],
+        'uploadDate' => '2020-03-19T21:12:41.000Z',
+        'contentUrl' => 'https://v19.muscdn.com/8bf11e3d42bab77f82d7a093d48cfed5/5ea3575d/video/tos/useast2a/tos-useast2a-ve-0068c003/2e16caf331794dd3bb95baf864210f7d/?a=1233&br=1912&bt=956&cr=0&cs=0&dr=0&ds=2&er=&l=20200424151648010115228069176F9653&lr=tiktok_m&qs=0&rc=M2x0Zjxrdmp4czMzNjczM0ApaGk1NWY4Ojw8N2Y7PDZpOWcwby5fc2ZlLW1fLS0uMTZzcy0zLjNjNC5hMWE1YjQ1YTE6Yw%3D%3D&vl=&vr=',
+        'embedUrl' => 'https://www.tiktok.com/embed/v2/6806030056956251397',
+        'commentCount' => '346',
+        'duration' => 'PT29S',
+        'audio' => [
+            'name' => 'Resistiré - Duo Dinamico',
+            'author' => 'Duo Dinamico',
+            'mainEntityOfPage' => [
+                '@type' => 'ItemPage',
+                '@id' => 'https://www.tiktok.com/music/Resistiré-250824169968246784'
+            ]
+        ],
+        'creator' => [
+            '@type' => 'Person',
+            'name' => 'Antena 3 Noticias',
+            'alternateName' => 'a3noticias',
+            'url' => 'https://www.tiktok.com/@a3noticias',
+            'interactionStatistic' => [
+                [
+                    '@type' => 'InteractionCounter',
+                    'interactionType' => [
+                        '@type' => 'http://schema.org/LikeAction'
+                    ],
+                    'userInteractionCount' => '139176'
+                ],
+                [
+                    '@type' => 'InteractionCounter',
+                    'interactionType' => [
+                        '@type' => 'http://schema.org/FollowAction'
+                    ],
+                    'userInteractionCount' => 21892
+                ]
+            ]
+        ],
+        'width' => 480,
+        'height' => 848,
+        'interactionStatistic' => [
+            [
+                '@type' => 'InteractionCounter',
+                'interactionType' => [
+                    '@type' => 'http://schema.org/WatchAction'
+                ],
+                'userInteractionCount' => 207568
+            ],
+            [
+                '@type' => 'InteractionCounter',
+                'interactionType' => [
+                    '@type' => 'http://schema.org/LikeAction'
+                ],
+                'userInteractionCount' => 10645
+            ],
+            [
+                '@type' => 'InteractionCounter',
+                'interactionType' => [
+                    '@type' => 'http://schema.org/ShareAction'
+                ],
+                'userInteractionCount' => 139
+            ]
+        ],
+        'mainEntityOfPage' => [
+            '@type' => 'ItemPage',
+            '@id' => 'https://www.tiktok.com/@a3noticias/video/6806030056956251397'
+        ]
+    ],
+    'oEmbed' => [
+        'version' => '1.0',
+        'type' => 'video',
+        'title' => 'Nuestro #AplausoSanitario más sincero. Para todos. Y recuerda, #QuédateEnCasa',
+        'author_url' => 'https://www.tiktok.com/@a3noticias',
+        'author_name' => 'Antena 3 Noticias',
+        'width' => '100%',
+        'height' => '100%',
+        'html' => '<blockquote class="tiktok-embed" cite="https://www.tiktok.com/@a3noticias/video/6806030056956251397" data-video-id="6806030056956251397" style="max-width: 605px;min-width: 325px;" > <section> <a target="_blank" title="@a3noticias" href="https://www.tiktok.com/@a3noticias">@a3noticias</a> <p>Nuestro <a title="aplausosanitario" target="_blank" href="https://www.tiktok.com/tag/aplausosanitario">#AplausoSanitario</a> más sincero. Para todos. Y recuerda, <a title="quédateencasa" target="_blank" href="https://www.tiktok.com/tag/quédateencasa">#QuédateEnCasa</a></p> <a target="_blank" title="♬ Resistiré - Duo Dinamico" href="https://www.tiktok.com/music/Resistiré-250824169968246784">♬ Resistiré - Duo Dinamico</a> </section> </blockquote> <script async src="https://www.tiktok.com/embed.js"></script>',
+        'thumbnail_width' => 480,
+        'thumbnail_height' => 848,
+        'thumbnail_url' => 'https://p16-va-default.akamaized.net/obj/tos-maliva-p-0068/26093075d77684fa3d60fd42bf50d48a',
+        'provider_url' => 'https://www.tiktok.com',
+        'provider_name' => 'TikTok'
+    ]
+];